[term] Move term filter func interface to term_int.h
